+++ lib/nerves/package.ex
@@ -104,9 +104,11 @@ defmodule Nerves.Package do
             end
           end)
       end
-    if ret == :ok do
-      Path.join(Artifact.dir(pkg, toolchain), @checksum)
-      |> File.write!(checksum(pkg))
+
+    case ret do
+      :ok -> Path.join(Artifact.dir(pkg, toolchain), @checksum)
+             |> File.write!(checksum(pkg))
+        _ -> :error
     end
   end
 
@@ -256,6 +258,7 @@ defmodule Nerves.Package do
         deps_path =
           File.cwd!
           |> Path.join(Mix.Project.config[:deps_path])
+          |> Path.expand
         if String.starts_with?(path, deps_path) do
           :hex
         else